a lot of changes

+++ rust2/grains/src/main.rs
@@ -1,7 +1,7 @@
 fn main() {
-    let mut result: u64 = 1;
+    let mut result: i128 = 1;
     let mut squares: u32 = 1;
-
+    
     result *= 2;
     while squares != 64 {
         print!(" Grains {result} ");
